prompt user for operator listener address
